tpl-lang \ v0.3.4 \ integer conversion & builtin module
1. All built-in functions separated to 'builtin.rs' module into
'BuiltIn' trait
2. `to_int8` full conversion support
